fix refactor helper functions into separate directory
