+++ helpers/index.ts
@@ -1,0 +1,7 @@
+export {
+  getFromLocalStorage,
+  setToLocalStorage,
+  removeFromLocalStorage,
+} from './localStorage';
+
+export {getAiResponse, getYoutubeSearchResults} from './dataFetch';
