+++ pages/index.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/naming-convention */
 import type {FormEvent} from 'react';
 import type {NextPage} from 'next';
 import Head from 'next/head';
@@ -13,76 +12,20 @@ import AiProfile from '../components/AiProfile';
 import WelcomeMessage from '../components/WelcomeMessage';
 import ShowResponse from '../components/ShowResponse';
 import UserPrompt from '../components/UserPrompt';
+import {
+  getFromLocalStorage,
+  setToLocalStorage,
+  removeFromLocalStorage,
+  getAiResponse,
+  getYoutubeSearchResults,
+} from '../helpers';
 
 import styles from './Home.module.css';
-
-const getFromLocalStorage = (key: string) => {
-  if (typeof window !== 'undefined') {
-    return window.localStorage.getItem(key);
-  }
-};
-
-const setToLocalStorage = (key: string, value: string) => {
-  if (typeof window !== 'undefined') {
-    return window.localStorage.setItem(key, value);
-  }
-};
-
-const removeFromLocalStorage = (key: string) => {
-  if (typeof window !== 'undefined') {
-    return window.localStorage.removeItem(key);
-  }
-};
 
 const Home: NextPage = () => {
   const messageEndRef = useRef<null | HTMLDivElement>(null);
   const [userInput, setUserInput] = useState('');
   const [results, setResults] = useState<ResultType[]>([]);
-
-  const getYoutubeSearchResults = async (song: string) => {
-    const youtubeApiKey = process.env.NEXT_PUBLIC_YOUTUBE_KEY;
-    const response = await fetch(
-      `https://www.googleapis.com/youtube/v3/search?part=snippet&maxResults=1&q=${song}&key=${youtubeApiKey}`,
-      {
-        method: 'GET',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-      },
-    );
-    const result = await response.json();
-    if (result.items) {
-      const resultVideo = result.items[0];
-      const resultVideoId = resultVideo.id.videoId;
-      const resultVideoThumbnail = resultVideo.snippet.thumbnails.high;
-      const resultVideoTitle = resultVideo.snippet.title;
-      const resultVideoObj = {
-        id: resultVideoId,
-        thumbnail: resultVideoThumbnail,
-        title: resultVideoTitle,
-      };
-      return resultVideoObj;
-    }
-
-    return {
-      id: 'error 403',
-      thumbnail: {url: 'nourl'},
-      title: 'error',
-    };
-  };
-
-  const getAiResponse = async (input: string) => {
-    const response = await fetch('/api/generate', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({prompt: input}),
-    });
-    const responseData = await response.json();
-
-    return responseData;
-  };
 
   const submitHandler = (input: string, event?: FormEvent<HTMLFormElement>) => {
     event?.preventDefault();
